Added support for MKS TFT32 version 1.0 and version 3.0
Support is now limited to using the original bootloader from MKS. Since the screen is initialized from the bootloader.
For V1. 0, support is limited to a small amount of flash memory. As a result, some of the images not used in this version of the screen were excluded from the distribution. For the same reason, font- word_unicode.fon is truncated to 208KB. At the moment, there were no errors in the interface related to these changes.